refactor(core): removed deprecated index exports
- move internal usage to `index-indernal`
- remove `index-deprecated`
