+++ packages/build-tools/src/process-url-dependencies.ts
@@ -1,6 +1,10 @@
 import type { UrlNode } from 'css-selector-tokenizer';
-import { isAsset, makeAbsolute, StylableMeta } from '@stylable/core';
-import { processDeclarationFunctions } from '@stylable/core/dist/process-declaration-functions';
+import type { StylableMeta } from '@stylable/core';
+import {
+    isAsset,
+    makeAbsolute,
+    processDeclarationFunctions,
+} from '@stylable/core/dist/index-internal';
 import { dirname } from 'path';
 
 function defaultFilter() {
