+++ packages/core/src/index-internal.ts
@@ -1,4 +1,53 @@
 export { safeParse } from './parser';
-export { process } from './stylable-processor';
-export { StylableTransformer } from './stylable-transformer';
+export { processorWarnings, process, StylableProcessor } from './stylable-processor';
+export {
+    StylableTransformer,
+    postProcessor,
+    replaceValueHook,
+    StylableExports,
+    transformerWarnings,
+    ResolvedElement,
+} from './stylable-transformer';
 export type { MappedStates } from './features';
+export { murmurhash3_32_gc } from './murmurhash';
+export { cssParse } from './parser';
+export type { OptimizeConfig, IStylableOptimizer, StateParsedValue } from './types';
+export {
+    nativeFunctionsDic,
+    nativePseudoClasses,
+    nativePseudoElements,
+    knownPseudoClassesWithNestedSelectors,
+} from './native-reserved-lists';
+export { fixRelativeUrls, collectAssets, isAsset, makeAbsolute } from './stylable-assets';
+export { emitDiagnostics, DiagnosticsMode, EmitDiagnosticsContext } from './report-diagnostic';
+export { StylableResolver, StylableResolverCache } from './stylable-resolver';
+export { CacheItem, FileProcessor, cachedProcessFile, processFn } from './cached-process-file';
+export { createStylableFileProcessor } from './create-stylable-processor';
+export { createDefaultResolver } from './module-resolver';
+export { packageNamespaceFactory } from './resolve-namespace-factories';
+import {
+    createBooleanStateClassName,
+    createStateWithParamClassName,
+    resolveStateParam,
+    stateErrors,
+    stateMiddleDelimiter,
+    booleanStateDelimiter,
+} from './pseudo-states';
+export const pseudoStates = {
+    createBooleanStateClassName,
+    createStateWithParamClassName,
+    resolveStateParam,
+    stateErrors,
+    stateMiddleDelimiter,
+    booleanStateDelimiter,
+};
+export { getFormatterArgs } from './helpers/value';
+export { getRuleScopeSelector } from './helpers/rule';
+export { BoxedValueArray, BoxedValueMap, createCustomValue } from './custom-values';
+export { expandCustomSelectors } from './stylable-utils';
+export { systemValidators } from './state-validators';
+export { processDeclarationFunctions } from './process-declaration-functions';
+export { reportDiagnostic } from './report-diagnostic';
+
+// remove
+export type { SRule } from './deprecated/postcss-ast-extension';
