+++ packages/rollup-plugin/src/plugin-utils.ts
@@ -1,4 +1,5 @@
-import type { Stylable, StylableExports, StylableMeta } from '@stylable/core';
+import type { Stylable, StylableMeta } from '@stylable/core';
+import type { StylableExports } from '@stylable/core/dist/index-internal';
 import type { PluginContext } from 'rollup';
 import type { StylableRollupPluginOptions } from './index';
 import { processUrlDependencies } from '@stylable/build-tools';
