+++ packages/schema-extract/src/main.ts
@@ -1,14 +1,16 @@
-import {
+import type {
     ClassSymbol,
     ElementSymbol,
     ImportSymbol,
-    cssParse,
-    StateParsedValue,
     StylableMeta,
-    StylableProcessor,
     VarSymbol,
 } from '@stylable/core';
-import type { MappedStates } from '@stylable/core/dist/index-internal';
+import {
+    MappedStates,
+    cssParse,
+    StateParsedValue,
+    StylableProcessor,
+} from '@stylable/core/dist/index-internal';
 import { getCssDocsForSymbol } from './cssdocs';
 import {
     MinimalPath,
